+++ example/storybook-templates/storybook-native/babel.config.js
@@ -1,0 +1,9 @@
+module.exports = function (api) {
+  api.cache(true);
+  return {
+    presets: ["babel-preset-expo"],
+    plugins: [
+      ["babel-plugin-react-docgen-typescript", { exclude: "node_modules" }],
+    ],
+  };
+};