+++ example/storybook/babel.config.js
@@ -1,5 +1,4 @@
 const path = require('path');
-console.log('herher');
 module.exports = function (api) {
   api.cache(true);
   return {
@@ -17,6 +16,10 @@ module.exports = function (api) {
                 '@gluestack-ui/config': path.join(
                   __dirname,
                   '../../packages/config/src/gluestack-ui.config'
+                ),
+                '@gluestack-ui/accordion': path.join(
+                  __dirname,
+                  '../../packages/unstyled/accordion/src'
                 ),
                 '@gluestack-style/react': path.join(
                   __dirname,
@@ -37,6 +40,10 @@ module.exports = function (api) {
                 '@react-native-aria/interactions': path.join(
                   __dirname,
                   '../../packages/react-native-aria/interactions/src'
+                ),
+                '@react-native-aria/accordion': path.join(
+                  __dirname,
+                  '../../packages/react-native-aria/accordion/src'
                 ),
                 '@react-native-aria/button': path.join(
                   __dirname,