+++ packages/config/src/theme/RadioLabel.ts
@@ -32,8 +32,8 @@ export const RadioLabel = createStyle({
     MozUserSelect: 'none',
     WebkitUserSelect: 'none',
     msUserSelect: 'none',
+    userSelect: 'none',
   },
-  'userSelect': 'none',
   '_dark': {
     'color': '$textDark400',
     ':checked': {