+++ packages/config/src/theme/SelectActionsheet.ts
@@ -5,4 +5,7 @@ export const SelectActionsheet = createStyle({
   height: '$full',
   justifyContent: 'flex-end',
   alignItems: 'center',
+  _web: {
+    pointerEvents: 'none',
+  },
 });