+++ packages/themed/src/components/Accordion/styled-components/Header.web.tsx
@@ -1,0 +1,5 @@
+import { H3 } from '@expo/html-elements';
+import { styled } from '@gluestack-style/react';
+export default styled(H3, {}, {
+  componentName: 'AccordionHeader',
+} as const);