+++ packages/styled/react/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @gluestack-style/react
 
+## 1.0.31
+
+### Patch Changes
+
+- fix: as forwarder conditional rendering
+
+## 1.0.30
+
+### Patch Changes
+
+- fix: ascomp issue
+
+## 1.0.29
+
+### Patch Changes
+
+- Fixed config reference issue
+- Fixed inline resolved stylesheet value
+
+## 1.0.28
+
+### Patch Changes
+
+- Exposed `GluestackStyleSheet` and `styleCSSIds` to plugins
+
+## 1.0.27
+
+### Patch Changes
+
+- Fixed plugin architecture (PR)[https://github.com/gluestack/gluestack-ui/pull/1490]
+
 ## 1.0.22
 
 ### Patch Changes