+++ packages/unstyled/accordion/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @gluestack-ui/accordion
 
+## 1.0.0
+
+### Major Changes
+
+- feat: stable release of accrodion component.
+
+## 0.0.3
+
+### Patch Changes
+
+- removed gluestack-style dependancy
+
 ## 0.0.2
 
 ### Patch Changes